[CLEAN] ft_p near completion

+++ srcs/client_cmd/client_cmd_ls.c
@@ -6,7 +6,7 @@
 /*   By: agrumbac <agrumbac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/18 19:47:09 by agrumbac          #+#    #+#             */
-/*   Updated: 2019/01/12 17:53:57 by agrumbac         ###   ########.fr       */
+/*   Updated: 2019/01/13 17:24:08 by agrumbac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 __attribute__((warn_unused_result))
 static bool		receive_body(int sock, t_ftp_header answer, char *body)
 {
-	int		ret;
+	ssize_t		ret;
 
 	if (answer.body_size > MAXPATHLEN)
 	{
@@ -50,7 +50,7 @@ bool			cmd_ls(int sock, char *client_input)
 	answer.type = ASW_MORE;
 	while (answer.type == ASW_MORE)
 	{
-		if (receive_answer(sock, &answer) == false) // a changer <= boucle
+		if (receive_answer(sock, &answer) == false)
 			return (false);
 		if (answer.type == ASW_OK || answer.type == ASW_MORE)
 		{
